src/test/java/com/appslandia/sweetsop/multipart/ReaderPartTest.java
package com.appslandia.sweetsop.multipart;

import java.io.ByteArrayOutputStream;
import java.io.IOException;
import java.io.Reader;
import java.io.StringReader;

import org.junit.Assert;
import org.junit.Test;

public class ReaderPartTest {

	@Test
	public void test() {
		ByteArrayOutputStream os = new ByteArrayOutputStream();

		Reader content = new StringReader("data");
		ReaderPart part = new ReaderPart("part", "part.dat", content, "text/plain");

		try {
			part.writeBody(os);
			String out = os.toString("UTF-8");

			Assert.assertEquals(out, "data");

		} catch (IOException ex) {
			Assert.fail(ex.getMessage());
		}
	}

	@Test
	public void test_null() {
		ByteArrayOutputStream os = new ByteArrayOutputStream();
		ReaderPart part = new ReaderPart("part", "part.dat", null, "text/plain");

		try {
			part.writeBody(os);
			byte[] out = os.toByteArray();

			Assert.assertEquals(part.getBodyLength(), 0);
			Assert.assertEquals(out.length, 0);

		} catch (IOException ex) {
			Assert.fail(ex.getMessage());
		}
	}
}
